requirements.txt
scikit-learn==0.20.1
sklearn==0.0
psycopg2-binary
django-heroku
gunicorn
requests